Run import and mock scripts in a transaction to avoid partially persisted data
